__is: accept unions over multiple tables that declare `__is` metamethod
Accordingly, accept `is` checks on table types that declare `__is`.

We disallow unions mixing tables with and without `__is`. This is a limitation
because to lift it we would have to implement code that transforms this

```
local u = R1 | R2 | {number} -- R1 and R2 have __is, {number} doesn't

if u is {number} then
   -- ...
end
```

into (effectively)

```
local u = R1 | R2 | {number} -- R1 and R2 have __is, {number} doesn't

if not (u is R1 or u is R2) then
   -- ...
end
```

In other words, "is" testing for the one table/userdata item without __is
would have to be defined in terms of the negation of the disjunction of
all the other cases.